o Integrated Qemu network user backend:
By default when adding any AnyEmuDevice image, the user network
backend will be set, allowing users to issue a DHCP request on the
last interface of the running emulated OS.
It can be turned off/on via the preferences.

o qemuprotocol -> 1

We will now use the default syntax by default, we are able to ship
the latest qemu.exe which uses this syntax.

o Fixed some typos in the code
o CONFIGSPECPATH for dynagen is updated for OpenBSD